chore: move send to embedded component
